# add RVV optimization for ZSTD_row_getMatchMask