This pull request introduces a RISC-V Vector (RVV) specific optimization for the `ZSTD_row_getMatchMask` function, replacing the generic SWAR implementation on RV64 platforms with V-extension support. The goal is to leverage RVV's parallel computation capabilities to improve performance on the RISC-V architecture. The implementation follows a "Construct-Select-Reduce" strategy, which is detailed in the code

## Performance
### Microbenchmark Results
A microbenchmark isolating the `ZSTD_row_getMatchMask` function shows a significant speedup compared to the SWAR fallback.

| rowEntries | Speedup   |
| ---------- | --------- |
| 16 bytes   | **3.45x** |
| 32 bytes   | **4.24x** |
| 64 bytes   | **4.73x** |

### Fullbench

The overall impact on the `fullbench` is modest. However, the new implementation shows a consistent small improvement and, most importantly, **no performance regression**.

Ten consecutive runs were performed for both the original version and the RVV-optimized version. The RVV version demonstrates a **~0.3%** average performance increase for this specific function under these test conditions.

| Version    | Average Speed        | Fastest Run          | Slowest Run      |
| ---------- | -------------------- | -------------------- | ---------------- |
| **Origin** | 1,081,809.8 MB/s     | 1,091,072.1 MB/s     | 1,068,435.0 MB/s |
| **RVV**    | **1,085,074.8 MB/s** | **1,096,995.4 MB/s** | 1,063,997.2 MB/s |


## Validation
- [x] All quick check pass (`make check`).  
- [x] All longer check pass（`make test`）
- [x] Static analysis reports no new issues (`make staticAnalyze`).